feat: Don't dereference common properties in the schemas
Effectively what needs to happen is that the `definitions` property for each schema is merged into a single property, and for the common schemas to be dereferenced into definitions.

We can take advantage of `JSON.stringify`'s replacer argument to repoint all `$refs` point at `common/<interface>.schema.json` with `#/definitions/<interface>`.